Create 01_Kallisto_scratch_NoNEED
The Kallisto is installed from scratch, and the reference for the human genome is also. To take the index, it requires a big RAM of about 8GB. But the group has the index, thus, no need to install and make. However, to do it from scratch, there are guidelines here.